pyproject.toml
[project]
name = "carevisitservice"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "dj-database-url (>=3.1.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-filter (>=25.2,<26.0)",
    "celery[redis] (>=5.6.2,<6.0.0)",
    "load-dotenv (>=0.1.0,<0.2.0)",
    "drf-yasg (>=1.21.14,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
